+++ saleor/payment/gateways/braintree/__init__.py
@@ -1,4 +1,3 @@
-import logging
 from typing import Dict, List
 
 import braintree as braintree_sdk
@@ -17,14 +16,7 @@ from ...interface import (
 from .errors import DEFAULT_ERROR_MESSAGE, BraintreeException
 from .forms import BraintreePaymentForm
 
-logger = logging.getLogger(__name__)
-
-
 # FIXME: Move to SiteSettings
-
-# If this option is checked, then one needs to authorize the amount paid
-# manually via the Braintree Dashboard
-CONFIRM_MANUALLY = False
 THREE_D_SECURE_REQUIRED = False
 
 # Error codes whitelist should be a dict of code: error_msg_override
@@ -38,22 +30,29 @@ ERROR_CODES_WHITELIST = {
 }
 
 
+def get_billing_data(payment_information: PaymentData) -> Dict:
+    billing = {}
+    if payment_information.billing:
+        billing_info = payment_information.billing
+        billing = {
+            "first_name": billing_info.first_name,
+            "last_name": billing_info.last_name,
+            "company": billing_info.company_name,
+            "postal_code": billing_info.postal_code,
+            "street_address": billing_info.street_address_1,
+            "extended_address": billing_info.street_address_2,
+            "locality": billing_info.city,
+            "region": billing_info.country_area,
+            "country_code_alpha2": billing_info.country,
+        }
+    return billing
+
+
 def get_customer_data(payment_information: PaymentData) -> Dict:
     """Provide customer info, use only for new customer creation."""
-    billing = payment_information.billing
     return {
         "order_id": payment_information.order_id,
-        "billing": {
-            "first_name": billing.first_name,
-            "last_name": billing.last_name,
-            "company": billing.company_name,
-            "postal_code": billing.postal_code,
-            "street_address": billing.street_address_1,
-            "extended_address": billing.street_address_2,
-            "locality": billing.city,
-            "region": billing.country_area,
-            "country_code_alpha2": billing.country,
-        },
+        "billing": get_billing_data(payment_information),
         "risk_data": {"customer_ip": payment_information.customer_ip_address or ""},
         "customer": {"email": payment_information.customer_email},
     }
@@ -94,7 +93,7 @@ def extract_gateway_response(braintree_result) -> Dict:
     }
 
 
-def create_form(data, payment_information, connection_params):
+def create_form(data, payment_information):
     return BraintreePaymentForm(data=data, payment_information=payment_information)